Change to .tsx file and add ESLint style
